moved ProductController into correct package
