+++ code/orders-api-rest/src/main/java/com/academy/orders/apirest/products/controller/ProductController.java
@@ -1,4 +1,4 @@
-package com.academy.orders.apirest.products;
+package com.academy.orders.apirest.products.controller;
 
 import com.academy.orders.apirest.products.mapper.ProductPreviewDTOMapper;
 import com.academy.orders.domain.product.usecase.GetAllProductsUseCase;
